Enable subsample filtering with augur filter's query argument
Adds support for a `query` key to individual subsampling definitions, allowing
users to filter strains by specifying the attribute values needed to retain
strains. This approach can make subsampling filters easier to read than the
corresponding exclusionary filters with `--exclude-where`.

Also adds an example of how to use the new `query` key in a county-level build
for King County in Washington State.

Note that the `--query` argument is only available in augur versions 8.0.0 and
later, so this commit also updates the required version of augur in the conda
environment file.